refactor: rename PascalCase components to kebab-case
