+++ src/components/auth/auth-required-modal.tsx
