+++ src/layouts/setting/setting-modal.tsx
@@ -2,7 +2,7 @@ import Analytics from '@/analytics'
 import { callEvent } from '@/common/utils/call-event'
 import Modal from '@/components/modal'
 import { type TabItem, TabManager } from '@/components/tab-manager'
-import { UpdateReleaseNotesModal } from '@/components/UpdateReleaseNotesModal'
+import { UpdateReleaseNotesModal } from '@/components/update-release-notes-modal'
 import { AboutUsTab } from './tabs/about-us/about-us'
 import { AppearanceSettingTab } from './tabs/appearance/appearance'
 import { GeneralSettingTab } from './tabs/general/general'
