+++ src/layouts/setting/tabs/account/tabs/rewards/components/referral-code-section.tsx
