+++ src/layouts/setting/tabs/account/tabs/rewards/rewardsTab.tsx
@@ -1,7 +1,7 @@
 import { ConfigKey } from '@/common/constant/config.key'
 import { SectionPanel } from '@/components/section-panel'
 import { useGetReferrals } from '@/services/hooks/user/referralsService.hook'
-import { ReferralCodeSection } from './components/ReferralCodeSection'
+import { ReferralCodeSection } from './components/referral-code-section'
 import { RewardTasks } from './components/tasks'
 import { RequireVerification } from '@/components/auth/require-verification'
 import { useAuth } from '@/context/auth.context'
