+++ src/layouts/setting/tabs/account/tabs/user-profile/user-profile.tsx
@@ -9,7 +9,7 @@ import {
 } from '@/services/hooks/user/userService.hook'
 import { AccountVerificationStatus } from '../../components/account-verification-status'
 import { ProfileDisplay } from '../../components/profile-display'
-import { ReferralCodeSection } from '../rewards/components/ReferralCodeSection'
+import { ReferralCodeSection } from '../rewards/components/referral-code-section'
 import { showToast } from '@/common/toast'
 import { translateError } from '@/common/utils/translate-error'
 import { ConfirmationModal } from '@/components/modal/confirmation-modal'
