+++ src/layouts/setting/tabs/general/components/select-city.tsx
@@ -5,7 +5,7 @@ import Modal from '@/components/modal'
 import { SectionPanel } from '@/components/section-panel'
 import { useGetCitiesList } from '@/services/hooks/cities/getCitiesList'
 import { useAuth } from '@/context/auth.context'
-import { AuthRequiredModal } from '@/components/auth/AuthRequiredModal'
+import { AuthRequiredModal } from '@/components/auth/auth-required-modal'
 import { useSetCity } from '@/services/hooks/user/userService.hook'
 import { TextInput } from '@/components/text-input'
 import { showToast } from '@/common/toast'
